+++ src/sender.h
@@ -18,7 +18,7 @@ public:
     virtual bool Initialize();
     virtual void Destroy();
     virtual bool Send(const Event::buf_t& buf);
-    virtual bool Send(const char* buf, int len);
+    virtual bool Send(const char* buf, size_t len);
     virtual const SendOption& GetOption() const;
     virtual std::string GetErrMsg(const std::string& sep = "\n") const;
     virtual std::string GetStatistics(int level = 1, const std::string& sep = "\n") const;
